+++ src/lib/components/theme-switch/ThemeSwitch.svelte
@@ -4,8 +4,9 @@
 
 	import WhiteBalanceSunny from 'svelte-material-icons/WhiteBalanceSunny.svelte'
 	import MoonWaningCrescent from 'svelte-material-icons/MoonWaningCrescent.svelte'
-  import type { Theme } from 'src/global';
-  import { Empty } from '$lib/components';
+	import type { Theme } from 'src/global'
+	import { Empty } from '$lib/components'
+  import { classMap } from '$helpers/classMap'
 
 	let isSystemDark = false
 
@@ -15,48 +16,64 @@
 		return 'light'
 	}
 
-  const Icon = (theme: Theme)  => getTheme(theme) === 'light' ? WhiteBalanceSunny : MoonWaningCrescent
+	const Icon = (theme: Theme) =>
+		getTheme(theme) === 'light' ? WhiteBalanceSunny : MoonWaningCrescent
 
-  let component = Empty
+	let component = Empty
 
-  onMount(() => {
-        const storageTheme = window.localStorage.getItem('theme') as Theme
-        if (storageTheme && 
-          (storageTheme.includes('light') || storageTheme.includes('dark'))
-        )
-          return updateThemeByStorage(storageTheme)
-        updateThemeOnInitBySystemPreferences(),
-        updateThemeOnSystemPreferencesChanges()
-    }
-  )
+	onMount(() => {
+		const storageTheme = window.localStorage.getItem('theme') as Theme
+		if (
+			storageTheme &&
+			(storageTheme.includes('light') || storageTheme.includes('dark'))
+		)
+			return updateThemeByStorage(storageTheme)
+		updateThemeOnInitBySystemPreferences(),
+			updateThemeOnSystemPreferencesChanges()
+	})
 
-  theme.subscribe((currentTheme) => component = Icon(currentTheme))
+	theme.subscribe((currentTheme) => (component = Icon(currentTheme)))
 
-  const updateThemeByStorage = (storageTheme: Theme) => {
-    theme.set(storageTheme)
-    component = Icon(storageTheme)
-  }
+	const updateThemeByStorage = (storageTheme: Theme) => {
+		theme.set(storageTheme)
+		component = Icon(storageTheme)
+	}
 
 	const handleOnClick = () => {
-    const newTheme = getTheme($theme) === 'light' ? 'dark' : 'light'
-    window.localStorage.setItem('theme', newTheme)
-    theme.set(newTheme)
-  }
+		const newTheme = getTheme($theme) === 'light' ? 'dark' : 'light'
+		window.localStorage.setItem('theme', newTheme)
+		theme.set(newTheme)
+	}
 
-  const updateThemeOnInitBySystemPreferences = () => {
-    isSystemDark = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches
-    component = Icon($theme)
-  }
+	const updateThemeOnInitBySystemPreferences = () => {
+		isSystemDark =
+			window.matchMedia &&
+			window.matchMedia('(prefers-color-scheme: dark)').matches
+		component = Icon($theme)
+	}
 
-  const updateThemeOnSystemPreferencesChanges = () => window.matchMedia('(prefers-color-scheme: dark)')
-    .addEventListener('change', event => {
-      isSystemDark = event.matches
-      component = Icon($theme)
-  })
+	const updateThemeOnSystemPreferencesChanges = () =>
+		window
+			.matchMedia('(prefers-color-scheme: dark)')
+			.addEventListener('change', (event) => {
+				isSystemDark = event.matches
+				component = Icon($theme)
+			})
 </script>
 
-<button class="icon" type="button" title="theme switch" on:click={handleOnClick} on:keyup={() => {}}>
-  <svelte:component this={component} />
+<button
+	class={classMap({
+    icon: true,
+    'dark-theme': $theme === 'dark',
+    'light-theme': $theme === 'light',
+  })}
+	type="button"
+	title="theme switch"
+	data-testid="theme-switch"
+	on:click={handleOnClick}
+	on:keyup={() => {}}
+>
+	<svelte:component this={component} />
 </button>
 
 <style lang="scss">