+++ src/routes/about-me/+page.svelte
@@ -3,7 +3,7 @@
 	import Gradient from '$lib/components/common/Gradient.svelte'
 	import { t } from '$lib/translations'
 
-  const resume = `Prazer, me chamo Ruan Augusto Pasta e atuo como desenvolvedor frontend a
+	const resume = `Prazer, me chamo Ruan Augusto Pasta e atuo como desenvolvedor frontend a
 		mais de 5 anos. Sou um amante do ecosistema JS e tudo o que o engloba, porem
 		estou sempre curioso para aprender novas tecnologias. Tenho bastante
 		interesse em me aprofundar mais no desenvolvimento de games com o uso de
@@ -12,7 +12,7 @@
 		Se quiser me conhecer um pouco mais ou ate mesmo conversar sobre alguma
 		proposta, estou aberto para conversa e para freelances seletos.`
 
-  let fragmentedResume = resume.split('.')
+	let fragmentedResume = resume.split('.')
 </script>
 
 <Card title={$t('common.menu.aboutMe')}>
@@ -24,16 +24,20 @@
 			rounded
 		>
 			<figure>
-				<img src="/images/ruanpasta.jpg" alt="Foto de Ruan Pasta" />
+				<img
+					src="/images/ruanpasta.jpg"
+					alt="Foto de Ruan Pasta"
+					data-testid="about-me-image"
+				/>
 			</figure>
 		</Gradient>
 	</div>
 
-  {#each fragmentedResume as text}
-    <p class="about-me-resume">
-      {text}
-    </p>
-  {/each}
+	{#each fragmentedResume as text}
+		<p class="about-me-resume">
+			{text}
+		</p>
+	{/each}
 </Card>
 
 <style lang="scss">