+++ tests/e2e/components/personalLinks.test.ts
@@ -1,0 +1,7 @@
+import { test } from '@playwright/test';
+
+test('should have personal links', async ({ page }) => {
+  await page.goto('/');
+
+  await page.innerText('.personal-links')
+});