+++ tests/e2e/pages/experience.test.ts
@@ -1,0 +1,13 @@
+import { test } from '@playwright/test'
+
+test.describe('works page', () => {
+	test.beforeEach(async ({ page }) => await page.goto('/experience'))
+
+	test('should have a card title', async ({ page }) => {
+		await page.innerText('.card__header__title')
+	})
+  
+	test('should have a accordion component', async ({ page }) => {
+    await page.innerText('.accordion')
+	})
+})