+++ vite.config.js
@@ -1,0 +1,15 @@
+import { sveltekit } from '@sveltejs/kit/vite';
+
+/** @type {import('vite').UserConfig} */
+const config = {
+  plugins: [sveltekit()],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    include: ['tests/unit/**/*.{test,spec}.ts'],
+    exclude: ['**/e2e/**'],
+    setupFiles: ['./setupTest.js']
+  }
+};
+
+export default config;