+++ src/lib/stores/common.ts
@@ -1,5 +1,23 @@
 import { writable } from 'svelte/store'
+import { browser } from '$app/environment'
+import { ThemeOptions } from '$/common'
+import type { Theme } from '$/global'
 
-const isMobile = writable<boolean>(false)
+// theme
+const getTheme = () => {
+	if (browser) {
+		return (localStorage.getItem('theme') as Theme) ?? ThemeOptions.System
+	}
+	return ThemeOptions.System
+}
 
-export { isMobile }
+const theme = writable<Theme>(getTheme())
+
+theme.subscribe((theme: Theme) => {
+	if (browser) return (localStorage.theme = theme)
+})
+
+// isMobile
+const isMobile = writable<boolean>(true)
+
+export { theme, isMobile }