+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-<!-- TODO: Update the changelog with issues links -->
+# 1.0.0
+[05/12/2022]
+
+- [5](https://github.com/ruanpasta/about-me/issues/5) - Create the custom line gradient component.
+- [6](https://github.com/ruanpasta/about-me/issues/6) - Create the Apresentation component.
+- [7](https://github.com/ruanpasta/about-me/issues/7) - Create Links on footer.
+- [9](https://github.com/ruanpasta/about-me/issues/9) - Create the button component.
+- [10](https://github.com/ruanpasta/about-me/issues/10) - Create Card component.
+- [13](https://github.com/ruanpasta/about-me/issues/13) - Create the skills page.
+- [11](https://github.com/ruanpasta/about-me/issues/11) - Use buttons component in Projects page.
+- [12](https://github.com/ruanpasta/about-me/issues/12) - Use Card component.
+- [15](https://github.com/ruanpasta/about-me/issues/15) - Create About me page.
+- [14](https://github.com/ruanpasta/about-me/issues/14) - Create the Accordion component.
+- [17](https://github.com/ruanpasta/about-me/issues/17) - Improve the Light theme.
+- [3](https://github.com/ruanpasta/about-me/issues/3) - Configure language switcher component.
+- [18](https://github.com/ruanpasta/about-me/issues/18) - Create some tests.
+- [19](https://github.com/ruanpasta/about-me/issues/19) - Create a desktop version.
+- [20](https://github.com/ruanpasta/about-me/issues/20) - Create the page switch component.
+- [2](https://github.com/ruanpasta/about-me/issues/2) - Configure automatic deployment with Netlify.
+- [23](https://github.com/ruanpasta/about-me/issues/23) - Improve mock texts.
+
 
 # 0.2.0
 [30/10/2022]